refactor(cli): split research import helpers
