+++ src/notebooklm/cli/research.py
@@ -14,13 +14,13 @@ from .helpers import (
     console,
     display_report,
     display_research_sources,
-    import_research_sources,
     json_output_response,
     require_notebook,
     resolve_notebook_id,
     with_client,
 )
 from .options import notebook_option
+from .research_import import import_research_sources
 
 # UI-only cap for the research summary preview shown in `research status` /
 # `research wait`. Unlike RPC error previews (see
