+++ tests/unit/cli/test_research.py
@@ -9,7 +9,7 @@ from notebooklm.notebooklm_cli import cli
 from .conftest import create_mock_client, patch_client_for_module
 
 research_module = importlib.import_module("notebooklm.cli.research")
-helpers_module = importlib.import_module("notebooklm.cli.helpers")
+research_import_module = importlib.import_module("notebooklm.cli.research_import")
 
 # =============================================================================
 # RESEARCH STATUS TESTS
@@ -179,7 +179,7 @@ class TestResearchWait:
         with (
             patch_client_for_module("research") as mock_client_cls,
             patch.object(
-                helpers_module, "import_with_retry", new_callable=AsyncMock
+                research_import_module, "import_with_retry", new_callable=AsyncMock
             ) as mock_import,
         ):
             mock_client = create_mock_client()
@@ -210,7 +210,7 @@ class TestResearchWait:
         with (
             patch_client_for_module("research") as mock_client_cls,
             patch.object(
-                helpers_module, "import_with_retry", new_callable=AsyncMock
+                research_import_module, "import_with_retry", new_callable=AsyncMock
             ) as mock_import,
         ):
             mock_client = create_mock_client()
@@ -277,7 +277,7 @@ class TestResearchWait:
         with (
             patch_client_for_module("research") as mock_client_cls,
             patch.object(
-                helpers_module, "import_with_retry", new_callable=AsyncMock
+                research_import_module, "import_with_retry", new_callable=AsyncMock
             ) as mock_import,
         ):
             mock_client = create_mock_client()
@@ -314,7 +314,7 @@ class TestResearchWait:
         with (
             patch_client_for_module("research") as mock_client_cls,
             patch.object(
-                helpers_module, "import_with_retry", new_callable=AsyncMock
+                research_import_module, "import_with_retry", new_callable=AsyncMock
             ) as mock_import,
         ):
             mock_client = create_mock_client()
